move freeStyle driver, pt 2
