+++ lib/core/device.js
@@ -28,7 +28,7 @@ var timeutils = require('../timeutils');
 
 var dexcomDriver = require('../drivers/dexcomDriver');
 var oneTouchMiniDriver = require('../oneTouchMiniDriver');
-var abbottFreeStyle = require('../abbottFreeStyle');
+var abbottFreeStyle = require('../drivers/abbottFreeStyle');
 
 var device = {
   log: require('../bows')('Device')
