+++ lib/drivers/abbottFreeStyle.js
@@ -18,7 +18,7 @@
 var _ = require('lodash');
 var async = require('async');
 var sundial = require('sundial');
-var struct = require('./struct.js')();
+var struct = require('../struct.js')();
 
 module.exports = function (config) {
   var cfg = _.clone(config);
